requirements.txt
numpy~=1.19.5
matplotlib~=3.4.2
Pillow~=8.3.1
fonttools~=4.26.2
future~=0.17.1
pydub~=0.23.1
opencv-python~=4.5.4.60
ffmpeg~=1.4